added the ability to remove and filter by the selected field and value

+++ src/main/java/org/app/service/BookService.java
@@ -1,10 +1,14 @@
 package org.app.service;
 
+import org.apache.log4j.Logger;
 import org.app.dto.Book;
+import org.app.exception.FilterOrRemoveByFieldException;
 import org.app.repository.ProjectRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.lang.reflect.Field;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
@@ -12,6 +16,7 @@ import java.util.Objects;
 public class BookService {
 
     private final ProjectRepository<Book> bookRepo;
+    private final Logger logger = Logger.getLogger(BookService.class);
 
     @Autowired
     public BookService(ProjectRepository<Book> bookRepo) {
@@ -30,11 +35,42 @@ public class BookService {
         }
     }
 
-    public void removeBookByField(String bookFieldToRemove, String bookFieldValueToRemove) {
-        bookRepo.removeItemByField(bookFieldToRemove, bookFieldValueToRemove);
+    public void removeBookByField(String bookFieldToRemove, String bookFieldValueToRemove) throws FilterOrRemoveByFieldException {
+        try {
+            for (Book book : bookRepo.retrieveAll()) {
+                Field bookField = book.getClass().getDeclaredField(bookFieldToRemove);
+                bookField.setAccessible(true);
+                Object bookValue = bookField.get(book);
+                if (Objects.nonNull(bookValue)) {
+                    if (bookValue.toString().equals(bookFieldValueToRemove)) {
+                        bookRepo.removeItemByField(bookFieldToRemove, bookFieldValueToRemove);
+                        logger.info("book is removed: " + book);
+                    }
+                }
+            }
+        } catch (NoSuchFieldException | IllegalAccessException ex) {
+            logger.info("remove book by field exception: " + ex.getMessage());
+            throw new FilterOrRemoveByFieldException("remove book by field exception: " + ex.getMessage());
+        }
     }
 
-    public List<Book> filterBooksByField(String bookFieldToFilter, String bookFieldValueToFilter) {
-        return bookRepo.filterItemsByField(bookFieldToFilter, bookFieldValueToFilter);
+    public List<Book> filterBooksByField(String bookFieldToFilter, String bookFieldValueToFilter) throws FilterOrRemoveByFieldException {
+        List<Book> filteredBooks = new ArrayList<>();
+        try {
+            for (Book book : bookRepo.retrieveAll()) {
+                Field bookField = book.getClass().getDeclaredField(bookFieldToFilter);
+                bookField.setAccessible(true);
+                Object fieldValue = bookField.get(book);
+                if (Objects.nonNull(fieldValue)) {
+                    if (fieldValue.toString().contains(bookFieldValueToFilter)) {
+                        filteredBooks.add(book);
+                    }
+                }
+            }
+        } catch (NoSuchFieldException | IllegalAccessException ex) {
+            logger.info("filter books by field exception: " + ex.getMessage());
+            throw new FilterOrRemoveByFieldException("filter books by field exception: " + ex.getMessage());
+        }
+        return filteredBooks;
     }
 }
